+++ src/types/globals.d.ts
@@ -2,6 +2,7 @@ interface PrunTile {
   id: string;
   frame: HTMLDivElement;
   anchor: HTMLDivElement;
+  docked: boolean;
   fullCommand: string;
   command: string;
   parameter: string | undefined;